requirements.txt
# Data Acquisition & Processing
datasets>=2.14.0
huggingface_hub>=0.16.0
transformers>=4.30.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0
boto3>=1.28.0              # For AWS S3 access
smart-open>=6.4.0          # For streaming from S3
chardet>=5.2.0             # For improved HTTP requests

# Preprocessing & Cleaning
regex>=2023.6.3
nltk>=3.8.1
scikit-learn>=1.3.0

# Tokenization & NLP
sentencepiece>=0.1.99

# Testing & Quality Assurance
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1        # For parallel test execution
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
coverage>=7.3.0
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0

# Visualization & Reporting
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.16.0

# Bias Analysis
scipy>=1.11.2              # For statistical tests
numpy>=1.24.0

# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1        # For parallel test execution
pytest-benchmark>=4.0.0     # For performance testing
pytest-timeout>=2.1.0      # For test timeouts
coverage>=7.3.0            # For code coverage reporting
faker>=19.3.0              # For generating test data

# Orchestration & Workflow
apache-airflow>=2.7.0

# Versioning & Schema Management
dvc>=3.30.0
great_expectations>=0.18.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Monitoring & Logging
prometheus-client>=0.17.0
grafana-api>=1.0.0
python-json-logger>=2.0.7

# Testing & Quality Assurance
pytest>=7.4.0
pytest-cov>=4.1.0

# Miscellaneous Utilities
matplotlib>=3.8.0
seaborn>=0.13.0

# For bias analysis and advanced ML
tensorflow>=2.13.0
fairlearn>=0.9.0

# FastAPI and Serving Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
mlflow>=2.7.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.24.0
safetensors>=0.4.0
psutil>=5.9.0
prometheus-client>=0.17.0
aiofiles>=23.2.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development and Testing for API
httpx>=0.25.0
requests>=2.31.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.22.0
factory-boy>=3.3.0

# Additional monitoring and observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
structlog>=23.2.0

# FastAPI and Serving Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
mlflow>=2.7.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.24.0
safetensors>=0.4.0
psutil>=5.9.0
prometheus-client>=0.17.0
aiofiles>=23.2.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

